replaced the working file with the new one
